Add first running version
This does not use Docker, as I decided we do not need it, and can make
due with run.sh and some config json files in
adv_symmetry/configs/*.json. Our profiles on the rippers should be
sufficiently isolated on their own.

This should have everything working and logging, and we should be able
to run this version as-is on the rippers to get any results wanted out.

The only thing I am still worried about is that the slope is not
processed correctly. The hypothesis was that symmetry would be *worse*
if the envoirement is asymetrical, but I think the current fitness
function does not test for that properly.